+++ common/src/main/java/net/mehvahdjukaar/moonlight/core/network/ClientBoundSendLoginPacket.java
@@ -23,15 +23,5 @@ public class ClientBoundSendLoginPacket implements Message {
     @Override
     public void handle(ChannelHandler.Context context) {
         AntiRepostWarning.run();
-
-        try {
-            //check on datapack registries on client
-            SoftFluidRegistry.empty();
-            BuiltInSoftFluids.WATER.value();
-        } catch (Exception e) {
-            throw new IllegalStateException("Not all required entries were found in datapack registry. How did this happen?" +
-                    "This is NOT a Moonlight issue. Do not report there. This can only be caused by some other mod messing up mod added datapack registries."+
-                    "Crashing now to prevent inevitable random crashes in-game.", e);
-        }
     }
 }